fix(entity-mapper): refactor entityComponents state
Before, the components were saved in a Map with key-value storage with
the key being a Symbol for the component. This caused problems when two
completely different components were the same at top level
(e.g. when using a HOC). Now, the components are saved in an array which
is searched wand matched on uuid & mapper (same logic as hnContext)