fix(Drupal page): export the DrupalPage.assureData method
In a previous version of hn-react, I could import { DrupalPage } from 'hn-react', but since the
release implementing the React context API the DrupalPage isn't exported as such anymore. Instead,
we now find a React.forwardRef component without the previous existing static method. To solve
this, the DrupalPage.assureData() should be exported again, so it is easy to upgrade my projects to
the latest version of hn-react.

Closes 52